// src/modules/inventory/reports/yarn-store/buyer-wise-yarn-possition-report/buyer-wise-yarn-possition-report-type.ts
export type IBuyerWiseYarnPossitionReport = {
    COMPANY_NAME: string;
    COMPANY_ADDRESS: string;
    BUYER_ID: string;
    BUYER: string;
    STYLE_ID: string;
    STYLENO: string;
    STYLENAME: string;
    PO_ID: string;
    PONO: string;
    DELIVERYDATE: string;
    YARN_ID: string;
    BOOKING_YARN: string;
    BOOKING_QTY: number;
    BOOKING_DATE: string;
    ALLOCATED_YARN: string;
    ALLOCATED_QTY: number;
    PI_DATE: string;
    PINO: string;
    PI_QUANTITY: number;
    BBLC_DATE: string;
    BBLC_NO: string;
    SUPPLIER: string;
    YARN_RECEIVED_DATE: string;
    TOTAL_AMOUNT: number;
    CURRENCY: string;
}